+++ workspaces/orchestrator/plugins/orchestrator-backend/src/service/api/mapping/V2Mappings.ts
@@ -144,6 +144,7 @@ export function mapToProcessInstanceDTO(
     duration: duration,
     // @ts-ignore
     workflowdata: variables?.workflowdata,
+    initiatorEntity: variables?.initiatorEntity as string,
     state: processInstance.state
       ? getProcessInstancesStatusDTOFromString(processInstance.state)
       : undefined,