+++ workspaces/adoption-insights/plugins/analytics-module-adoption-insights/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @red-hat-developer-hub/backstage-plugin-analytics-module-adoption-insights
 
+## 0.1.0
+
+### Minor Changes
+
+- a66e0b0: Backstage version bump to v1.36.1
+
 ## 0.0.2
 
 ### Patch Changes