+++ workspaces/orchestrator/plugins/orchestrator-common/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## @red-hat-developer-hub/backstage-plugin-orchestrator-common [1.13.1](https://github.com/janus-idp/backstage-plugins/compare/@red-hat-developer-hub/backstage-plugin-orchestrator-common@1.13.0...@red-hat-developer-hub/backstage-plugin-orchestrator-common@1.13.1) (2024-08-02)
 
+## 1.27.2
+
+### Patch Changes
+
+- c6b54ad: Updated dependency `@openapitools/openapi-generator-cli` to `2.20.2`.
+- 7f6ca8a: add filter by variables and nested variables
+
 ## 1.27.1
 
 ### Patch Changes