+++ workspaces/orchestrator/plugins/orchestrator/CHANGELOG.md
@@ -1,5 +1,13 @@
 ### Dependencies
 
+## 3.0.2
+
+### Patch Changes
+
+- Updated dependencies [c6b54ad]
+- Updated dependencies [7f6ca8a]
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-common@1.27.2
+
 ## 3.0.1
 
 ### Patch Changes